Fix: View was overlapping deck details on iOS
Using padding made the view clickable on iOS. And since the view
overlaps the positioned absolute deck details, none of the buttons
worked.

When switching to using margin for padding, it worked, but as margin
doesn't increase the view height, we needed to measure the deck info
height and add that to the scrollTo function to compensate.